weatherbot/requirements.txt
requests
azure-functions
